Updated unit test structure to ensure that each file has a corresponding unit test
