--- src/decko/helper/parallel.py
@@ -1,3 +1,0 @@
-from concurrent.futures import ThreadPoolExecutor
-
-
