+++ tests/helper/test_exceptions.py
