+++ tests/helper/test_properties.py
@@ -1,0 +1,13 @@
+import pytest
+from src.decko.app import Decko
+
+dk = Decko(__name__)
+
+
+def test_registration():
+    """
+    Basic use case for testing whether a custom function was successfully registered
+    :return:
+    :rtype:
+    """
+    pass
